test: relocate MySQL frame tests
